encoderStats moved to its own headerfile
