+++ tests/suite/CMakeLists.txt
@@ -1,5 +1,7 @@
 include_directories(${CMAKE_SOURCE_DIR}/common)
+include_directories(${CMAKE_SOURCE_DIR}/common/os)
 include_directories(${CMAKE_SOURCE_DIR}/unix)
+include_directories(${CMAKE_SOURCE_DIR}/unix/common)
 include_directories(${X11_INCLUDE_DIR})
 include_directories(${FLTK_INCLUDE_DIR})
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON CACHE INTERNAL "")
@@ -10,6 +12,13 @@ target_sources(suite PRIVATE
   Manager.cxx
   Server.cxx
   Client.cxx
+  TestWindow.cxx
+  ../../vncviewer/PlatformPixelBuffer.cxx
+  ../../vncviewer/Surface.cxx
+  ../../vncviewer/Surface_X11.cxx
+  #../../unix/xserver/hw/vnc/XserverDesktop.cc
+  #../../unix/xserver/hw/vnc/vncBlockHandler.c
+  #../../common/os/os.cxx
   benchmark/Benchmark.cxx
   codec/PPMDecoder.cxx
   codec/PNGDecoder.cxx
@@ -26,13 +35,19 @@ target_sources(suite PRIVATE
   codec/timed/TimedTightJPEGEncoder.cxx
   codec/timed/TimedZRLEEncoder.cxx
   codec/timed/timedEncoderFactory.cxx
+  codec/timed/encoderStats.h
   io/FrameInStream.cxx
   streams/DummyInStream.cxx
   streams/DummyOutStream.cxx
 )
 
-target_link_libraries(suite rfb)
+target_link_libraries(suite rfb ${FLTK_LIBRARIES})
 target_compile_features(suite PRIVATE cxx_std_11)
+
+find_package(OpenMP)
+if(OpenMP_CXX_FOUND)
+    target_link_libraries(MyTarget PUBLIC OpenMP::OpenMP_CXX)
+endif()
 
 # Recording tool
 add_subdirectory(recorder)
